Introduce .usize() for casting convenience
When we use `as`, we usually expect the numeric value to be preserved
and don't expect to get a large unsigned number when casting from a
negative number, for example.

Add a utility trait for convenience and build time verification of what
we usually want when we say `as usize`. This is similar to
[`.to_bits()`](https://doc.rust-lang.org/std/primitive.pointer.html#method.to_bits)
for raw pointers.

Cleanup some warnings from `cargo docs` while we are at it.