Port gen_direct_jump() and gen_block_series()
With the advent of destructors I was able to simplify
`gen_block_series()` a lot. Also taking a risk by batching FFI calls for
furture simplification. I mentioned this deviation in a comment.

I moved the limit_block_version() and the block allocation into
gen_single_block() like in the C code to keep the logic for
`gen_block_series()` concise. I hope this is fine with you @maximecb?

This is now hitting:

    thread '<unnamed>' panicked at 'assertion failed: !(branch.shape == target_branch_shape && block.start_addr != branch.end_addr)', src/core.rs:1291:13

Which I haven't figured out.