Fix sweeping cursor movement bug
The compaction algorithm relies on the try_move method returning false
to know when movement has failed and it needs to increment the sweeping
cursor and try again.

A bug in this function meant that we also returned false when the object
being moved was not moveable (eg. it had a finaliser). In this case the
function would return false, the sweeping page would be incremented and
then we'd try - and fail - again.

So the first time we hit an unmoveable object the sweeping cursor would
be effectively pushed all the way to the compact cursor and compaction
would stop - leaving many uncompacted pages.