Include Rust YJIT in libruby_static.a
Prior to this bigdecimal had linking issue which it detected during
mkmf configuration and that made it not build properly. This let's us
try to boot `make test-all`.

This works by combining two archive libraries with ar(1). I also tried a
couple of other approaches. `cargo rustc -- --emit obj` outputs an
object file but at a path that's hard to find and considered internal to
Cargo according to the manual. `ld -r --whole-archive` seems to be an
alternative, but unfortunately is unavailable with the linker on macOS.