See how CI reacts when we don't merge archive libs
Patch from ianks on GitHub from discussions on the upstreaming PR. I'm
pretty sure this won't work, but if it does it'd be a nice
simplification.

The original PR where I introduced the static lib merging is
https://github.com/Shopify/ruby/pull/248
At the time, we weren't running the `make test-all` on CI, so the errors
weren't on GitHub.