Fix test_bug_reporter_add when JITS are enabled
Commit 31b2cd38c5dcf4a0c51ca56ecdddf7461b8ac86c changed the way Ruby
description works to be initialised earlier in the boot process and more
accurately reflect whether a JIT was enabled.

As part of this change some crash description handling was removed from
`test_bug_reporter.rb`. Removing this handling causes the test to fail
when run with JITs enabled.

This is because the test uses `assert_in_out_err` which uses `spawn` to
execute a new Ruby process. The args passed to this process do not
include `--mjit` or `--yjit` when required. Previously this test worked
around that by modifying the matching regex to remove the `/+[MY]JIT/`
match.

That special handling code was removed as part of 31b2cd.

This commit reinstates some special case code for when the tests are run
with a JIT enabled, but instead of mutating the Regex, we'll tell the
spawned process to also use the same JIT that's enabled on it's parent.