Only autocompact on majors when it's done by object allocations
I suspect that we're doing a major by malloc, and that malloc is
happening when the geniv table (or some hash table) is being resized.
If we're resizing a hash that contains Ruby references, and we need to
update the references during compaction, then we have a problem because
we can't be "mid reallocate" *and* updating references at the same time.
The reallocated hash may not have the right references inside