Remove ractor_core.h dependancy on rb_gc_obj_slot_size
This commit has had to introduce another function that returns the
address at the end of the slot. This is only needed when
RACTOR_CHECK_MODE > 0 because some debug information is stored in
between slots.

I like this approach better however as we're asking the GC for
information about the object itself, rather than exposing information
about the heap slots to the ractor.

If the underlying implementation changed, asking the GC for the end of
an allocated region still could make sense, whereas relying on it
exposing a slot size might not.