[Experimental] Implement RubyVM::YJIT.pause
This allows to pause and resume YJIT compilation at any point.

Note that it won't prevent YJIT from compiling when it hits a stub.

But this is flexible enough to allow to prevent compilation on a
per code section basis (for single threaded apps). e.g. per controller
in a Rails app.