Mark cc->cme_ for refinement callcaches as well
This seems to be required for the same reason that it's required for
super (added in 36023d5cb751d62fca0c27901c07527b20170f4d). I'm not
exactly sure _how_ this is possible though; surely the CME must be
strongly held by the refinement module?

Hopefully the bugtracker can help me shed some light on the reasons for
this. In any case, this _feels_ right and fixes the crash:

[Fixes #19994]