Arg nodes need special treatment if not being passed to a method
The arguments of `ReturnNode`, `NextNode`, `BreakNode`, etc. will always build an array out of the arguments. This is not the case if the arguments are passed to a method, but that case is already handled by `pm_setup_args`.

For all other case, we need to handle the different kinds of arguments and make sure we build an array, when more then one argument is passed.