Enable Ruby to dynamically link against a custom GC library
We made it so Ruby builds and links against a librubygc so we can move gc
routines into a shared library.

High level goal is to:

(on Linux)
$ LD_PRELOAD=custom_gc.so ruby -e'1'

(on macOS)
$ DYLD_LIBRARY_PATH=other_path/ ruby -e'1'

The dyld man page recommends that when we want to replace a shared library, we
manipulate `DYLD_LIBRARY_PATH`. This means that when we link librubygc, we need
to use a _relative path_ for linking.

Most people don't have the Ruby install path in DYLD_LIBRARY_PATH.

So by default, Ruby uses an absolute path when linking libruby. This is a problem
for the librubygc use case because no manipulation of DYLD_LIBRARY_PATH
(or DYLD_INSERT_LIBRARIES) will allow us to override librubygc. If we use _relative_
linking, then simply manipulating DYLD_LIBRARY_PATH will allow us to test out a
new GC as a dynamic library.

We think the default for linking librubygc should be relative.

With relative linking, we're able to successfully replace the librubygc dynamic
library on macOS:

Build Ruby like this:

```
./autogen.sh
./configure --disable-install-doc --disable-shared --enable-shared-gc
make -j
```

Custom GC:

```
 #include <stdio.h>
 #include <stdlib.h>

void
GC_Init(void) {
    fprintf(stderr, "===== >>> Loading our custom GC file\n");
}
```

Compile with:

```
clang -shared -dynamiclib rubygc2.c -o custom/librubygc.3.4.dylib
```

Ruby without custom GC:

```
$ ./ruby -e' '
Loading our GC file
```

Ruby with custom GC:

```
$ DYLD_LIBRARY_PATH=./custom ./ruby -e' '
===== >>> Loading our custom GC file
```

Miniruby is statically linked so you can't replace it's GC. This only works with Ruby.