Ensure windows has access to LIBRUBYGC
We aren't going to implement a shared object for windows (yet) but we
still need to statically build the file correctly and provide the needed
values.

Some of these can proably be deleted but since I don't have a windows
machine it's easier to just make sure everything is defined and remove
them later.

Hoping this fixes the following error from Windows / Visual Studio

```
../src/common.mk(455) : fatal error U1083: target macro '$(LIBRUBYGC_A)' expands to nothing
```