Refactor ary_heap_free_ptr
This PR:

* `ary_heap_free_ptr` took an `ary` but it wasn't used for anything so
I deleted the argument.
* `ary_heap_free` was static and just calls `ary_heap_free_ptr` so
update callers and delete that function.

Motivation:

We're currently exploring what a GC API looks like. Through that process
we're looking at MMTK and how that's implemented / what functions it
changes in order to use it's own GC. While doing that work I noticed
these methods and feel the indirection is unnecessary as it's not
public, nor does it really make the code that much cleaner.