Implement String#append_bytes(String)
[Feature #20594]

A handy method to construct a string out of multiple chunks.

Contrary to `String#concat`, it only takes a single argument
and doesn't do any encoding negociation. The string content is
appended, and the receiver encoding is preserved.

It's the caller responsibility to check for `String#valid_encoding?`
in cases where it's needed.