Support MMTK_HEAP_MODE, MMTK_HEAP_MIN and MMTK_HEAP_MAX
MMTK_HEAP_MODE configures whether the heap is in fixed size mode or
dynamic. Dynamic is default - setting this variable to "fixed" will
switch it to fixed mode, anything else is ignored.

MMTK_HEAP_MIN configures the minimum size of the heap, can use SSI
units, so 1GiB etc. This is ignored when the heap is in fixed mode.

MMTK_HEAP_MAX configures the maximum size of the heap, can use SSI
units, so 1GiB etc. This is valid in both fixed and dynamic heap modes.

By default the dynamic heap mode will use up to 80% of the available
physical memory of the machine, and 1Mb minimum.

If MMTK_HEAP_MODE is set to fixed but without MMTK_HEAP_MAX, then a
default of 100Mb will be used.