Add a small buffer for allocating fstrings
This commit adds a small mmap'd buffer that we can allocate fstrings
from. The buffer is read-only by default, when we allocate a string we
unlock the buffer, memcopy in to the buffer, then lock the buffer again.

We're hoping that this will catch unexpected writes in to the middle of
a string.

Co-Authored-By: Alan Wu <XrXr@users.noreply.github.com>