Remove mark maybe from `imemo_tmpbuf`
If `xcalloc` is used instead of `xmalloc` for `imemo_tmpbuf`, then
calling `rb_gc_mark_locations` on the internals of the `imemo_tmpbuf` is
no longer necessary.

The main goal is to avoid mark maybe / mark and pin on objects where it
isn't actually required. This is required for MMTk moving immix
implementation, however it's more efficient in the standard GC as well
to avoid pinning unnecessarily. Otherwise objects are pinned and not
compacted.

[link issue]

Co-authored-by: Aaron Patterson <tenderlove@ruby-lang.org>