[WIP] Use rb_field_get when transitioning too too complex
This is an attempt to fix a bug in CI which I cannot adequately explain
yet.

Instead of using rb_obj_id during the too complex table evacuation,
which risks triggering a shape transition we can use rb_field_get to
directly grab the existing object id (we know it exists already), and
then manually make sure the id_to_obj_tbl is updated if it has been
created already.