Get rid of FL_EXIVAR
The `shape_id` already provide most of that information.
If we know the object isn't T_OBJECT/T_CLASS then we can simply
check if it has the root shape and isn't complex, which can be
done with a single mask.

In a few cases, we don't know the type of the object and in addition
have to test it, but it's a very small overhead.