make methods on ENV safe if they load an encoding
Most of the time this should work fine, but if a native extension calls
`setlocale(SOME_LOCALE)`, and that encoding is not loaded, it will load
next time `rb_locale_encoding()` is called. This is called for methods on ENV
like ENV#[]. Since these ENV methods acquire the VM lock, we need to
make sure the encoding is loaded prior to locking it.

Calling `setlocale()` like this is supported. The documentation for
`rb_locale_encoding()` states:

```
 * This  is dynamic.   If  you  change the  process'  locale  by e.g.   calling
 * `setlocale(3)`, that should also change the return value of this function.
 *
 * There is no official way for Ruby scripts to manipulate locales, though.
 rb_encoding *rb_locale_encoding(void);
```